+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/feed/rabbitmq/configurations/DynamicRabbitMQDefinitionRegistrar.java
@@ -80,7 +80,7 @@ public class DynamicRabbitMQDefinitionRegistrar implements BeanDefinitionRegistr
             factory.setConcurrentConsumers(rabbitConnectionConfiguration.getConcurrentConsumers());
             factory.setMaxConcurrentConsumers(rabbitConnectionConfiguration.getMaxConcurrentConsumers());
             factory.setPrefetchCount(rabbitConnectionConfiguration.getPrefetchCount());
-            factory.setRecoveryInterval(rabbitConnectionConfiguration.getNetworkRecoveryInterval().longValue());
+            factory.setRecoveryInterval(rabbitConnectionConfiguration.getNetworkRecoveryInterval());
             factory.setMissingQueuesFatal(false);
             return factory;
         });